+++ .github/actions/setup/action.yml
@@ -16,7 +16,6 @@ runs:
           cargo risczero install
           cargo risczero --version
 
-      # TODO: https://github.com/InfinityVM/InfinityVM/issues/120
       - name: Install sp1 tooling
         shell: bash
         run: |
