+++ crates/zkvm/Cargo.toml
@@ -14,7 +14,6 @@ thiserror = { workspace = true }
 risc0-zkvm = { workspace = true }
 risc0-binfmt = { workspace = true }
 
-# TODO: https://github.com/InfinityVM/InfinityVM/issues/120
 sp1-sdk = { workspace = true }
 
 [dev-dependencies]
