+++ programs/risc0/mock-consumer/methods/guest/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "mock-consumer-guest"
+name = "mock-consumer-risc0-guest"
 version = "0.1.0"
 edition = "2021"
