+++ programs/sp1/mock-consumer/Cargo.toml
@@ -10,5 +10,5 @@ path = "src/lib.rs"
 
 [build-dependencies]
 sp1-helper = { workspace = true }
-sp1-build = "3.1.0"
+sp1-build = { workspace = true }
