+++ programs/sp1/mock-consumer/src/lib.rs
@@ -1,3 +1,3 @@
 /// The ELF (executable and linkable format) file for the Succinct RISC-V zkVM.
-pub const SP1_MOCK_CONSUMER_GUEST_ELF: &[u8] =
+pub const SP1_MOCK_CONSUMER_RISC0_GUEST_ELF: &[u8] =
     include_bytes!("../../../../elf/riscv32im-succinct-zkvm-elf");
